Add harness HTML and JS
Now we need a way to run our code and the easiest way to do this is to simply run it in a browser. Let's create `index.html` containing the harness `HTML`, a single `script` tag that includes `bower_components/requirejs/require.js` with a `data-main` attribute set to `index.js`.

In `index.js` we'll keep our `requirejs` configuration. Let's keep it simple for now and just set the configuration option  for `baseUrl` to `bower_components`, add a dependency to `mu-emitter/main` and a `callback` handler that will receive `Emitter` as its only argument.

You should now be able to load `index.html` in browser of choice and pop open the developer console with a single log line with `Emitter` on it.